stack trace updates:
limited support on Alpha - in general case, even with frame pointer,
stack trace on alpha is impossible without looking at the symbol table
frame pointer does get saved on the stack, but you have no idea where
and where the return address is saved. So the best we can do without
the symbol table is look for magic opcodes and try to guess how big 
each frame is and where the return address was hidden from the 
instruction parameters. In practice, we can usually go up 3-4 frames 
before we hit some nasty frame that the current code cannot figure
out. This is actually not too bad, especially when we already have the query

Also cleaned up messages, print more variables, tell the user of
how much memory mysqld could potentially use, and warn of
what can happen with default STACK_SIZE and a lot of connections if
coredump happens when there are more than 200 connections.