Bug#17926: mysql.exe crashes when ctrl-c is pressed in windows
SIGINT is handled in funny ways on windows, which could lead to problems when
Control-C was pressed in the client during a long-running query.  Now Control-C
during a query aborts that query (by sending KILL to the server on a second
connexion), while Control-C outside of a running query terminates the client.