Bug#21543: 5.0.24 breaks ABI compatibility for python bindings: \ 	InterfaceError on connect
Removed the bool flag from the st_mysql_options struct, since it adds 
another word in size to the memory size and shifts member memory locations
down, both of which break binary-interface compatibility.

Instead, use a flag, 2**30, in the client_options bit-field to represent
that the client should check the SSL certificate of the server.