BUG#9678: Client library hangs after network communication failure           (back-port to 4.0)
Socket timeouts in client library were used only on Windows.
Additionally, in 4.0 write operations erroneously set read timeout.

The solution is to use socket timeouts in client library on all
systems were they are supported, and to differentiate between read
and write timeouts.

No test case is provided because it is impossible to simulate network
failure in current test suite.