mysql-test/mysql-test-run.pl  +  mysql-test/mysql-test-run.sh
Necessary changes if one of the test scripts is to be used with a RPM installation (bug#17194).

This change handles finding the server and the other programs,
but it does not solve the problem to get a writable "var" directory.
If we want to avoid world-writable directories below "/usr/share/mysql-test" (and we do!), 
any automatic solution would require fixed decisions which may not match the local installation.

For the Perl script, use "--vardir"; for the shell script, create "mysql-test/var" manually.