bug #22634 (partitioned tables with DATA DIRECTORY option corrupts table)
In fact the problem is inside the ARCHIVE engine.
It stores real datapath inside the ARM file, but doesn't implement
specific ha_archive::rename_table function.
As the ALTER TABLE statement first creates the table with the temporary
name and then renames it to the normal one, we get the temporary name
saved in ARM what leads to "can't open file" error.
Code modified to store only path in the ARM, without filename.