WL#3561 - transactional LOCK TABLE WL 3594 - transactional LOCK TABLE Testing
Added a new clause to the LOCK TABLE statement:
LOCK TABLE table_name IN { SHARE | EXCLUSIVE } MODE [NOWAIT]

This allows for transactional locks.
- They do _not_ commit a transaction.
- One can access tables not mentioned in the LOCK TABLE statement.
- One can add more tables to the locked set.
- Tables are unlocked at the end of transaction.

If one or more of
- transactional and standard locks are mixed in a statement,
- non-transactional (standard READ/WRITE) locks exist already,
- or a storage engine does not support transactional locking,
apply, the locks are converted to standard (READ/WRITE) locks
with warnings. In strict mode the conversion is rejected with
an error message.